# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: ghidra-rest
  version: "v1"
  summary: REST API in front of Ghidra's headless analyzer.
  description: |
    Upload a binary, poll the job, read back functions, decompiled C, strings,
    symbols, imports, exports, types, cross references and raw bytes as JSON.

    One job is one `analyzeHeadless` run that exports a complete artifact set
    and exits. Result endpoints only read those files, so they answer `409`
    until the job reaches `done`.

    Addresses accept `401000`, `0x401000`, `00401000` and `RAM:00401000`; they
    normalise to the same key (lower case, no `0x`, no leading zeros, address
    space prefix kept).
  license:
    name: See repository
...
servers:
  - url: http://127.0.0.1:8080
    description: Local container or binary

# Both schemes are the same shared token. Applied globally; it is enforced only
# when GHIDRAREST_API_TOKEN is set, and /healthz opts out entirely.
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /v1/results/{id}/signatures:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      tags: [signatures]
      operationId: signatures
      summary: Prototypes edited on this job
      description: |
        Also reports whether this job can be retyped at all, and which calling
        convention names its compiler spec defines.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignaturesResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/NotDone" }
  /v1/results/{id}/function/{addr}/signature:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - $ref: "#/components/parameters/Addr"
    put:
      tags: [signatures]
      operationId: setSignature
      summary: Set a function prototype
      description: |
        Applies a C prototype to the function and re-decompiles it together with
        every caller. The function name inside the prototype is ignored --
        renaming is a separate concern.

        `calling_convention` is deliberately separate from the prototype text:
        Ghidra's C parser accepts `__cdecl` inside a signature and then discards
        it, leaving locked parameter storage with an `unknown` convention, which
        is what makes the decompiler print
        `Unknown calling convention ... parameter storage is locked`.

        PUT rather than POST because setting a prototype is idempotent. Runs are
        serialised across the whole server; each is a JVM.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prototype]
              properties:
                prototype:
                  type: string
                  description: "C, single line. The same text Ghidra's Edit Function Signature takes. A '*' written against the function name is moved onto the return type before parsing, so long *f(void) and long * f(void) behave identically."
                  example: long make_secret(byte *secret)
                calling_convention:
                  type: string
                  description: |
                    Optional. Must be one of `calling_conventions` from
                    `GET /v1/results/{id}/signatures`. Omit to leave it alone.
                  example: __stdcall
      responses:
        "200":
          description: Applied
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignatureApplied" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No function at that address
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: |
            Job is not `done`, or it kept no Ghidra project and so cannot be
            retyped at all.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: |
            The prototype did not parse, or the calling convention is not one
            this program defines. Nothing was applied: the transaction is rolled
            back whole.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignatureRejected" }
    delete:
      tags: [signatures]
      operationId: clearSignature
      summary: Restore the pre-edit prototype
      description: |
        Re-applies what the analyser said before the first edit and drops the
        ledger row. A restore, not an undo: Ghidra has no undo across processes,
        so the types come back but the restored signature is stored as
        `USER_DEFINED`.
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignatureApplied" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: No signature has been set on that address
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409": { $ref: "#/components/responses/NotDone" }
        "422":
          description: The recorded original no longer applies
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignatureRejected" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>`. Enforced only when
        `GHIDRAREST_API_TOKEN` is set; unset means the API is fully open.
        Comparison is constant time.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: The same shared token, for clients that prefer a custom header.
  parameters:
    JobId:
      name: id
      in: path
      required: true
      description: Job id, 24 hex characters.
      schema: { type: string, pattern: "^[0-9a-f]{24}$" }
      examples:
        ls: { value: "346ab74fe7fee34191cc8adb" }
    Addr:
      name: addr
      in: path
      required: true
      description: |
        Address in any accepted form: `401000`, `0x401000`, `00401000`,
        `RAM:00401000`.
      schema: { type: string }
      examples:
        plain: { value: "104000" }
        prefixed: { value: "0x00104000" }
        space: { value: "external:1" }
...
  responses:
...
    BadRequest:
      description: Malformed body, bad base64, missing `file` part, or empty upload
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: A token is configured and the request did not carry it
      headers:
        WWW-Authenticate:
          schema: { type: string, examples: ['Bearer realm="ghidra-rest"'] }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such job
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotDone:
      description: The job has not reached `done`, so it has no results
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      description: Every error in this API has this shape.
      required: [error, status]
      properties:
        error: { type: string }
        status: { type: integer }
      examples:
        - { error: "job is running, results are not available", status: 409 }
...
    Parameter:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        ordinal: { type: integer }
    CallEdge:
      type: object
      properties:
        address: { type: string }
        name: { type: string }
    Function:
      type: object
      properties:
        address: { type: string, description: Normalised key. }
        address_display: { type: string, description: Ghidra's own rendering. }
        name: { type: string }
        namespace: { type: string }
        signature: { type: string }
        calling_convention: { type: string }
        return_type: { type: string }
        size: { type: integer }
        parameter_count: { type: integer }
        is_thunk: { type: boolean }
        is_external: { type: boolean }
        is_inline: { type: boolean }
        has_varargs: { type: boolean }
        no_return: { type: boolean }
        stack_frame_size: { type: integer }
        parameters: { type: array, items: { $ref: "#/components/schemas/Parameter" } }
        calls: { type: array, items: { $ref: "#/components/schemas/CallEdge" } }
        called_by: { type: array, items: { $ref: "#/components/schemas/CallEdge" } }
    SignatureEntry:
      type: object
      description: One prototype guttex-style clients have pushed into Ghidra.
      properties:
        address: { type: string, example: "10135e" }
        prototype: { type: string, example: "long make_secret(byte * secret)" }
        calling_convention: { type: string, example: "__stdcall" }
        original:
          type: string
          description: |
            What the analyser said before the first edit. A reset re-applies
            this, which is the only recovery available -- Ghidra has no undo
            across processes.
          example: "undefined make_secret(void)"
        original_calling_convention: { type: string, example: "unknown" }
        at: { type: string, format: date-time }
    SignaturesResponse:
      type: object
      properties:
        job: { type: string }
        editable:
          type: boolean
          description: false when this job kept no Ghidra project; nothing here can be retyped.
        count: { type: integer }
        signature:
          type: array
          items: { $ref: "#/components/schemas/SignatureEntry" }
        calling_conventions:
          type: array
          description: |
            Names this program's compiler spec defines, and the only ones a PUT
            will accept. Empty on analyses predating the field that have never
            been edited.
          items: { type: string }
          example: ["MSABI", "__stdcall", "__thiscall", "processEntry", "syscall"]
    SignatureApplied:
      type: object
      properties:
        job: { type: string }
        address: { type: string }
        ok: { type: boolean, enum: [true] }
        before: { type: string, example: "undefined make_secret(void)" }
        prototype: { type: string, example: "long make_secret(byte * secret)" }
        calling_convention: { type: string }
        original: { type: string }
        set_at: { type: string, format: date-time }
        function:
          allOf:
            - $ref: "#/components/schemas/Function"
          description: The functions.json entry as it now reads.
        redecompiled:
          type: array
          description: |
            Everything whose `/decompile` response just changed: the retyped
            function plus every caller, because a caller's C names the callee's
            parameters and consumes its return value.
          items: { type: string }
          example: ["10135e", "1013d0"]
        duration_ms: { type: integer, example: 21344 }
    SignatureRejected:
      type: object
      description: The request was well formed and the prototype was not.
      properties:
        error:
          type: string
          example: 'this program has no calling convention "__cdecl"; it accepts MSABI, __stdcall, __thiscall, processEntry, syscall'
        status: { type: integer, enum: [422] }
        address: { type: string }
        before: { type: string }
        prototype: { type: string }
        duration_ms: { type: integer }
...